refactor(commands): extraer CommandResult/emit compartido y derivar tipo de collectDeclaredOrchestrators
